refactor: 상품 수량에 대한 책임 분리를 위해 Quantity 추가
